SQL FIle
Java code for SQL is used to connect Java programs to databases and execute SQL commands to manipulate the data in those databases. In a Java SQL program, the first step is to establish a connection to the database using a JDBC driver. The driver is responsible for communicating with the database and translating Java calls into SQL commands that the database can understand, which are all written in the code